AI_AVATAR_THAKUMARJHULI/requirements.txt
Flask==3.0.0
gTTS==2.5.0
requests==2.31.0
python-dotenv==1.0.0
